Rename the reducer index to index.js
